+++ src/server/src/KNTC.Domain/FileAttachments/FileAttachment.cs
@@ -28,7 +28,7 @@ public class FileAttachment : FullAuditedEntity<Guid>
         SetTenTaiLieu(tenTaiLieu);
     }
     public Guid IdHoSo { get; set; }
-    public short GiaiDoan { get; set; }
+    public int GiaiDoan { get; set; }
     public string TenTaiLieu { get; private set; }
     public int HinhThuc { get; set; }
     public DateTime ThoiGianBanHanh { get; set; }
